# decision-agent/agent-backend/agent-factory/cmd/openapi-docs/generated/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Agent Factory 智能体工厂 API 文档",
        "title": "Agent Factory API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:30777",
    "basePath": "/api/agent-factory",
    "paths": {
...
        "/v1/observability/agent/{agent_id}/conversation/{conversation_id}/session": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定对话的session列表，包含session的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "可观测性"
                ],
                "summary": "获取指定对话的session列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent_id",
                        "name": "agent_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "conversation_id",
                        "name": "conversation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回session列表",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/v1/observability/agent/{agent_id}/conversation/{conversation_id}/session/{session_id}/run": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定对话的session下的run列表，返回每个run的详情列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "可观测性"
                ],
                "summary": "获取指定session下的run列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent_id",
                        "name": "agent_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "conversation_id",
                        "name": "conversation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "session_id",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回run列表",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/v3/published/agent-tpl": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "已发布模板列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "已发布",
                    "模板"
                ],
                "summary": "已发布模板列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "- 分页marker（用于获取下一页数据） - base64编码的json字符串 - base64编码前的json格式： ``` { \\",
                        "name": "pagination_marker_str",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "403": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer {token} - 需要有效的 OAuth 访问令牌",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}